Simplify the test files
