+++ visual-tests/tests/.empty-test-template.js
@@ -1,18 +1,7 @@
-import { test, expect } from '@playwright/test';
-import path from 'path';
+import { test } from '../imports/test-runner';
 import { helpers } from '../imports/helpers';
 
-const stylesToAdd = [
-  helpers.cssFiles.cookieInfo,
-  helpers.cssFiles.dynamicDataFreeze
-];
-
-test(helpers.testTitle(path.basename(__filename)), async({ page }, workerInfo) => {
-  helpers.init(workerInfo);
-  await page.goto(helpers.testURL);
-  await expect(page).toHaveTitle(helpers.expectedPageTitle);
-  stylesToAdd.forEach(item => page.addStyleTag({ path: helpers.cssPath(item) }));
-
+test(__filename, async({ page }) => {
   const table = page.locator(helpers.selectors.mainTable);
 
   await table.waitFor();
@@ -21,5 +10,4 @@ test(helpers.testTitle(path.basename(__filename)), async({ page }, workerInfo) =
   // const thead = table.locator(helpers.selectors.mainTableHead);
 
   /* your test here */
-
 });
