+++ visual-tests/tests/change-rows-order.spec.js
@@ -1,21 +1,11 @@
-import { test, expect } from '@playwright/test';
-import path from 'path';
+import { test } from '../imports/test-runner';
 import { helpers } from '../imports/helpers';
 
-const stylesToAdd = [
-  helpers.cssFiles.cookieInfo,
-  helpers.cssFiles.dynamicDataFreeze
-];
-
-test(helpers.testTitle(path.basename(__filename)), async({ page }, workerInfo) => {
-  helpers.init(workerInfo);
-  await page.goto(helpers.testURL);
-  await expect(page).toHaveTitle(helpers.expectedPageTitle);
-  stylesToAdd.forEach(item => page.addStyleTag({ path: helpers.cssPath(item) }));
-
+test(__filename, async({ page }) => {
   const table = page.locator(helpers.selectors.mainTable);
 
   await table.waitFor();
+
   const mainTableFirstColumn = table.locator(helpers.selectors.mainTableFirstColumn);
   const cell = mainTableFirstColumn.locator(helpers.findCell({ row: 4, cell: 1, cellType: 'th' }));
